Move XIO code to xlib/xio.go and xlib/xioterm.go
